shared: fix/explicity ignore a few size-related assumptions
For the rv_array sizing, we already assume that the size of the rv_array
is not going to exceed INT_MAX, let alone what SIZE_MAX likely is on the
platform, so hopefully we're not comparing to something outside of that
range.

+++ src/client/headers/shared/containers.hpp
@@ -1414,7 +1414,7 @@ namespace intercept::types {
                 }
             }
             reserve(n_);
-            if (n_ > base::_n) {  //adding elements, need to default init
+            if (static_cast<int>(n_) > base::_n) {  //adding elements, need to default init
                 for (size_t i = base::_n; i < n_; ++i) {
 #pragma warning(suppress : 26409)
                     ::new (base::_data + i) Type();
